backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
Pillow==11.1.0
djangorestframework-simplejwt
psycopg2-binary==2.9.10
drf-spectacular==0.29.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
